+++ CHANGELOG.md
@@ -7,6 +7,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## Changed
 - Support for new, flexible wire protocol (V4).
 
+### Added
+- added new method in NoSQLHandle queryIterable to return the results of a
+query in an iterable/iterator format. The returned QueryIterableResult should
+be used in a try-with-resources statement to ensure proper closing of
+resources.
+- updated NoSQLHandle interface to extend AutoClosable
+
 ## [5.3.2] 2022-03-21
 
 ### Added